third_party/WebKit/LayoutTests/media/track/text-track-cue-exceptions.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test.js"></script>
</head>
<body>
    <script>
        description("Verify that TextTrackCue exceptions are properly messaged to developers.");

        var cue = new VTTCue(0, 0, "Test.");

        function testInfinityAndNaN(property) {
            shouldThrow("cue." + property + " = Number.NaN;");
            shouldThrow("cue." + property + " = Number.POSITIVE_INFINITY;");
            shouldThrow("cue." + property + " = Number.NEGATIVE_INFINITY;");
        }

        function testPercentage(property) {
            shouldThrow("cue." + property + " = -1;");
            shouldThrow("cue." + property + " = 101;");
        }

        testInfinityAndNaN("startTime");
        testInfinityAndNaN("endTime");

        shouldThrow("cue.vertical = 'Invalid!';");

        cue.snapToLines = false;
        testPercentage("line");

        testPercentage("position");
        testPercentage("size");

        shouldThrow("cue.align = 'Invalid!';");
    </script>
</body>
</html>
